# Implement multiple dispatch