--- a/aten/src/ATen/core/ATenDispatch.h
+++ b/aten/src/ATen/core/ATenDispatch.h
@@ -60,6 +60,8 @@ class CAFFE2_API ATenOpTable {
     function_table_[static_cast<int64_t>(tid)] = fn;
   }
 
+  void* getFallbackOp(TensorTypeId tid) const;
+
   void* getOp(TensorTypeId tid) const {
     // You might think we can minorly optimize this further by maintaining a
     // bitmask of registered operator keys, so we don't select dispatch ids
@@ -67,9 +69,7 @@ class CAFFE2_API ATenOpTable {
     // get a Variable CPUTensor, if there is no variable registration, you'll
     // fall back to the CPU implementation.  Is this what you want?  Unlikely...
     if (function_table_[static_cast<int64_t>(tid)] == nullptr) {
-      TORCH_CHECK(function_table_[static_cast<int64_t>(TensorTypeId::UndefinedTensorId)] != nullptr,
-          "No function is registered for schema ", schema_, " on tensor type ", toString(tid));
-      return function_table_[static_cast<int64_t>(TensorTypeId::UndefinedTensorId)];
+      return getFallbackOp(tid);
     }
     return function_table_[static_cast<int64_t>(tid)];
   }