@suo 2019-09-19
This broke master I think: 

```
Sep 19 21:44:52 RuntimeError: Source files: ['TensorMethods.h'] did not match generated files.  To update the source files, set environment variable GEN_TO_SOURCE or run "cp /var/lib/jenkins/workspace/build/aten/src/ATen/core_tmp/TensorMethods.h /var/lib/jenkins/workspace/cmake/../aten/src/ATen/core"

```